Moves Token into separate file.
