+++ src/lexer/Token.cpp
@@ -1,0 +1,10 @@
+/**
+ * @file Token.cpp
+ * @author Edward Palmer
+ * @date 2025-01-14
+ *
+ * @copyright Copyright (c) 2025
+ *
+ */
+
+#include "Token.hpp"
